# services/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: SMS SMP Negeri API
  description: "Kontrak tunggal web + mobile. Acuan baca: docs/07-api-contract.md"
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Login semua device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken: { type: string }
                      user: { type: object }
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Tukar refresh cookie dengan pasangan token baru (rotasi)
      responses:
        '200': { description: OK }
        '401': { description: Sesi berakhir }
...
  /percakapan/{id}/pesan:
    get:
      tags: [fase2]
      summary: Histori pesan percakapan (WALI_KELAS, ORANG_TUA)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { description: OK } }
    post:
      tags: [fase2]
      summary: Kirim pesan dalam percakapan (WALI_KELAS, ORANG_TUA)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '201': { description: Terkirim } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT